Implement correct routeing for Go+Angular
Implement correct gongular routing using supported locales and
serving the associated index.html frontend file.

In case the requested locale or path was not found, an error 404
is returned.
